refactor: Split SwiftUI views up into separate files
We had a massive ContentView. Used codex for these prompts.

Prompt 1:
This Xcode SwiftUI app currently has one massive ContentView.swift file. Please carefully review it and plan the
refactoring steps that are needed to break it up into separate files per SwiftUI view. If you have questions, ask. Review
your work afterwords and see if things can be improved or cleaned up. Make sure the build doesn't break.

Prompt 2:
Go ahead and implement the plan

Prompt 3:
Please look at the AGENTS.md file and update it with the relevant changes you've made in the refactor

Prompt 4:
Please add concise top-level doc comments to all public structs and classes.

Prompt 5:
This looks great! Please make one more improvement: Create an appropriate Group in the Xcode project to put all SwiftUI
views into expect for the top-level PerceptronVizApp.
