+++ PerceptronViz/SafeCollectionAccess.swift
@@ -1,0 +1,13 @@
+//
+//  SafeCollectionAccess.swift
+//  PerceptronViz
+//
+//  Created by Johannes Fahrenkrug on 18.09.25.
+//  https://springenwerk.com
+//
+
+extension Array {
+    subscript(safe index: Index) -> Element? {
+        indices.contains(index) ? self[index] : nil
+    }
+}
